[Java] Fix NPE when closing nodes

+++ src/main/java/uk/co/real_logic/benchmarks/rtt/aeron/LiveReplayMessageTransceiver.java
@@ -102,7 +102,10 @@ public final class LiveReplayMessageTransceiver extends MessageTransceiver
         if (ownsArchiveClient)
         {
             closeAll(aeronArchive, mediaDriver);
-            mediaDriver.context().deleteAeronDirectory();
+            if (null != mediaDriver)
+            {
+                mediaDriver.context().deleteAeronDirectory();
+            }
         }
     }
